hooks: add post-receive hook to build Sphinx doc
The hook will build the documentation if the Git config key
'documentation.source' is set. It expects the Sphinx documentation
source to be in the path defined by the value of that key, relative to
the root of the Git repository. It will build the documentation in
"/var/www/doc/${projectid}/html".

The Redmine Embedded plugin can then be configured to display static
content from that directory by setting its HTML directory to
'/var/www/doc/{PROJECT}/html/'.